Move component to new directory
