+++ frontend/src/components/PinnedBoardsEntry/PinnedBoardsEntry.styled.js
